requirements.txt
grpcio==1.0.4
grpc-google-cloud-speech-v1beta1==0.14.0
six==1.10.0
requests==2.12.4
google-auth==0.5.0
PyAudio==0.2.9enum34==1.1.6
futures==3.0.5
google-auth==0.5.0
googleapis-common-protos==1.5.0
grpc-google-cloud-speech-v1beta1==0.14.0
grpcio==1.0.4
httplib2==0.9.2
oauth2client==3.0.0
playsound==1.2.1
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyAudio==0.2.9
requests==2.12.4
rsa==3.4.2
six==1.10.0
wheel==0.24.0
